fixed validations bugs

+++ fitness/db/migrate/20101120002134_create_workout_sessions.rb
@@ -6,7 +6,7 @@ class CreateWorkoutSessions < ActiveRecord::Migration
 	  t.integer :event_id
 	  t.integer :exercise_id
 
-	  t.integer :estimated_mins
+	  t.float :estimated_mins
 
 
       t.timestamps
